+++ workspaces/backend/src/config/settings/prod.py
@@ -1,0 +1,8 @@
+from .base import *  # noqa: F403
+from .base import env
+
+DEBUG = False
+ALLOWED_HOSTS = env.list("ALLOWED_HOSTS", default=[])
+SECURE_PROXY_SSL_HEADER = ("HTTP_X_FORWARDED_PROTO", "https")
+SESSION_COOKIE_SECURE = True
+CSRF_COOKIE_SECURE = True